% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Test for effect of SNP on outcome variance using the LAD-BF model}
\usage{
model(data, x, y, covar1 = NULL, covar2 = NULL)
}
\arguments{
\item{data}{Dataframe of observations}

\item{x}{Name of SNP dosage}

\item{y}{Name of outcome}

\item{covar1}{Optional vector of covariates to include in the first-stage model}

\item{covar2}{Optional vector of covariates to include in the second-stage model}
}
\value{
res Vector of results: var(Y|G==0), var(Y|G==1), var(Y|G==2) and test P-value
}
\description{
Test for effect of SNP on outcome variance using the LAD-BF model
}
