% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_bs}
\alias{model_bs}
\title{Bootstrap function to obtain SE for LAD-BF effects}
\usage{
model_bs(data, indices, x, y, covar1 = NULL, covar2 = NULL)
}
\arguments{
\item{data}{Dataframe of observations}

\item{indices}{Vector of nrow indices for bootstrapping SEs}

\item{x}{Name of SNP dosage}

\item{y}{Name of outcome}

\item{covar1}{Optional vector of covariates to include in the first-stage model}

\item{covar2}{Optional vector of covariates to include in the second-stage model}
}
\value{
result Vector of betas
}
\description{
Usage result <- boot::boot(data=data, statistic=varGWASR::model_bs, R=500, x="SNP", y="outcome", covar1=c("c1", "c2"), covar2=c("c3", "c4")) \%>\% broom::tidy
}
